+++ src/cmd/magix.js
@@ -2,6 +2,8 @@
     author:xinglie.lkf@taobao.com
  */
 define('magix', ['$'], function(require) {
+    if (typeof DEBUG == 'undefined') DEBUG = true;
+
     var $ = require('$');
     var G_NOOP = function() {};
     /*#if(modules.hasDefaultView){#*/
@@ -77,8 +79,11 @@ define('magix', ['$'], function(require) {
             target: node
         });
     };
+
+    Inc('../tmpl/safeguard');
     Inc('../tmpl/magix');
     Inc('../tmpl/event');
+    Inc('../tmpl/state');
     /*#if(modules.router){#*/
     var Router_Edge;
     var G_IsFunction = $.isFunction;
@@ -92,9 +97,10 @@ define('magix', ['$'], function(require) {
             Router_WinLoc.hash = path;
         }
     };
-    var Router_Update = function(path, params, loc, replace, lQuery) {
+    var Router_Update = function(path, params, loc, replace, silent, lQuery) {
         path = G_ToUri(path, params, lQuery);
         if (path != loc.srcHash) {
+            Router_Silent = silent;
             Router_UpdateHash(path, replace);
         }
     };
@@ -187,9 +193,10 @@ define('magix', ['$'], function(require) {
             WinHistory[replace ? 'replaceState' : 'pushState'](G_NULL, G_NULL, path);
         };
         var Router_Popstate;
-        var Router_Update = function(path, params, loc, replace) {
+        var Router_Update = function(path, params, loc, replace, silent) {
             path = G_ToUri(path, params);
             if (path != loc.srcQuery) {
+                Router_Silent = silent;
                 Router_UpdateState(path, replace);
                 if (Router_Popstate) {
                     Router_Popstate(1);
@@ -309,7 +316,6 @@ define('magix', ['$'], function(require) {
     Inc('../tmpl/tmpl');
     Inc('../tmpl/partial');
     Inc('../tmpl/updater');
-
     Inc('../tmpl/view');
     /*#if(modules.service){#*/
     var G_Type = $.type;